+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 2.0.0
+
+### Updates
+
+* Replace `className` parameter by `selectorString` to use `Element.matches` function
+* `nodeName` parameter can be a single string or an Array
+* Update dependencies
+* Update tests
+
 ## 1.0.1
 
 ### Updates